Feat: Add support for creating secrets
In order to enable creating secrets from the cli much like you edit an
existing secret, this commit adds a new subcommand along with an
implementation. It updates the QuerySecrets trait and adds a real
implementation for aws to actually create the secret. And finally, it
adds a new error when we can't properly create the secret.