Feat: Add support for deleting secrets
In order to enable deleting secrets from the cli this commit adds
a new subcommand along with an implementation. It updates the
QuerySecrets trait and adds a real implementation for aws to actually
delete the secret. And finally, it adds a new error when we can't
properly delete the secret.